+++ src/main/scala/optics/poly/impl/GetOneImpl.scala
@@ -1,0 +1,32 @@
+package optics.poly
+
+
+class GetOneImpl[-S, +A](val get: S => A) extends GetterImpl[GetOne, S, A]: 
+  self => 
+
+  final def getOption: S => Option[A] = 
+    s => Some(get(s))
+
+  final def getAll: S => List[A] = 
+    s => List(get(s))
+
+  override def preComposeGetMany[S0](impl1: GetManyImpl[S0, S]): GetManyImpl[S0, A] = 
+    GetManyImpl(s0 => impl1.getAll(s0).map(get))
+
+  // override def preComposeGetOneOrMore[S0](impl1: GetOneOrMoreImpl[S0,S]): GetOneOrMoreImpl[S0, A] = 
+  //   GetOneOrMoreImpl(s0 => impl1.getOneOrMore(s0).map(get))
+
+  override def preComposeGetOption[AllowedByBoth >: (GetOne | GetOption) <: OpticCan, S0](impl1: GetOptionImpl[S0, S]): GetOptionImpl[S0, A] = 
+    GetOptionImpl(s0 => impl1.getOption(s0).map(get))
+
+  override def preComposeGetOne[S0](impl1: GetOneImpl[S0, S]): GetOneImpl[S0, A] = 
+    GetOneImpl(s0 => get(impl1.get(s0)))
+
+  def andThen[ThatCan <: OpticCan, C](impl2: GetterImpl[ThatCan, A, C]): GetterImpl[GetOne | ThatCan, S, C] = 
+    impl2.preComposeGetOne(this)
+
+  override def doGet(using GetOne <:< GetOne): S => A = get
+  override def doGetOption(using GetOne <:< GetOption): S => Option[A] = getOption
+  override def doGetAll(using GetOne <:< GetMany): S => List[A] = getAll
+
+end GetOneImpl
